refactor: adjust folder structure
