+++ data/recipes.ts
@@ -1,4 +1,4 @@
-import { Recipe } from "@/common/types"
+import { Recipe } from "@/types"
 
 export const recipes: Recipe[] = [
   {
